requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
google-generativeai>=0.3.0
pinecone-client>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.3.0
numpy>=1.24.3
transformers>=4.32.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.6.1
pytest>=7.4.0
httpx>=0.24.1
python-multipart>=0.0.6
pymongo
langchain>=0.2.1
langchain-community>=0.2.1
boto3
python-multipart
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.2
-e .
